+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.5] - 2021-07-03
+
+### Added
+
+- Added Main Seeder
+- Added Service Provider Trait
+
+### Changed
+
+- Seeders and Service Providers refactored
+
+### Fixed
+
 ## [1.1.4] - 2021-07-02
 
 ### Added